layer.conf: set preferred ELE firmware version
There is a potential bug with the firmware-ele-imx recipe in
meta-imx-frdm, in which the check for file existence is incomplete.
Moreover, it uses an older version of the firmware (1.3) versus 2.2.
The explicit set for preferred ELE version provides both transparency
and allow us to use the recipe provided by meta-freescale, which does
not contain this bug.

SFL: #19235
Change-Id: I75292ecaa230d0296a96a84569e5806c2e1351fe